+++ TIDALDL-PY/tidal_dl/paths.py
@@ -151,10 +151,9 @@ def getTrackPath(track, stream, album=None, playlist=None):
     # artist
     artists = __fixPath__(TIDAL_API.getArtistsName(track.artists))
     artistID = __fixPath__(TIDAL_API.getArtistsID(track.artists))
-
     primary_artist = getattr(track, 'artist', None)
-    trackArtistID = __tokenValue__(getattr(primary_artist, 'id', None))
-    trackArtistName = __tokenValue__(getattr(primary_artist, 'name', None))
+    trackArtistID = __fixPath__(__tokenValue__(getattr(primary_artist, 'id', None)))
+    trackArtistName = __fixPath__(__tokenValue__(getattr(primary_artist, 'name', None)))
 
     # title
     title = __fixPath__(track.title)
@@ -176,7 +175,8 @@ def getTrackPath(track, stream, album=None, playlist=None):
         retpath = SETTINGS.getDefaultPathFormat(Type.Track)
     hasStreamIdentifier = __hasStreamIdentifierToken__(retpath)
     retpath = retpath.replace(R"{TrackNumber}", number)
-    retpath = retpath.replace(R"{ArtistName}", artists)
+    retpath = retpath.replace(R"{ArtistName}", trackArtistName)
+    retpath = retpath.replace(R"{ArtistsName}", artists)
     retpath = retpath.replace(R"{ArtistID}", artistID)
     retpath = retpath.replace(R"{TrackArtistID}", trackArtistID)
     retpath = retpath.replace(R"{TrackArtistName}", trackArtistName)
@@ -208,12 +208,10 @@ def getVideoPath(video, album=None, playlist=None):
 
     # get artist
     artists = __fixPath__(TIDAL_API.getArtistsName(video.artists))
-    artist = __fixPath__(video.artist.name) if video.artist is not None else ""
     artistID = __fixPath__(TIDAL_API.getArtistsID(video.artists))
-
     primary_artist = getattr(video, 'artist', None)
-    videoArtistID = __tokenValue__(getattr(primary_artist, 'id', None))
-    videoArtistName = __tokenValue__(getattr(primary_artist, 'name', None))
+    videoArtistID = __fixPath__(__tokenValue__(getattr(primary_artist, 'id', None)))
+    videoArtistName = __fixPath__(__tokenValue__(getattr(primary_artist, 'name', None)))
 
     # explicit
     explicit = "(Explicit)" if video.explicit else ''
@@ -227,8 +225,9 @@ def getVideoPath(video, album=None, playlist=None):
     if retpath is None or len(retpath) <= 0:
         retpath = SETTINGS.getDefaultPathFormat(Type.Video)
     retpath = retpath.replace(R"{VideoNumber}", number)
-    retpath = retpath.replace(R"{ArtistName}", artists)
+    retpath = retpath.replace(R"{ArtistName}", videoArtistName)
     retpath = retpath.replace(R"{ArtistID}", artistID)
+    retpath = retpath.replace(R"{ArtistsName}", artists)
     retpath = retpath.replace(R"{VideoArtistID}", videoArtistID)
     retpath = retpath.replace(R"{VideoArtistName}", videoArtistName)
     retpath = retpath.replace(R"{VideoTitle}", title)